Close value
Cosmin's passion for video games led him to start programming so he could create games for other people. After launching a few simple apps on the Play Store, he built a multiplayer game.

The dynamics of the game are as follows:

Each player is initially dealt a card with a string of n natural numbers
To win, the player must correctly identify the PAIR of each number in the string on their playing card.
A pair of an x-value in the string on a card is a number y that is part of the same string and has the following property:

if x is even, y is the minimum of the numbers greater than x to the left of x in the string
if x is in an odd position, y is the minimum of the numbers greater than x to the right of x in the string
Input data
On the first line you will find a natural number n, representing the total number of values ​​on a card, and on the next line a string of n natural numbers, representing the elements of the string of numbers on that card.

Output data
The program will display a string of n numbers on the screen, representing the string of pairs corresponding to each element x in the string. If there is a number x that has no such pair in the string, the value -1 will be displayed instead of its pair.

Restrictions and clarifications
0 < n ≤ 100
0 ≤ x ≤ 1000
Example
Input data Output data
7
24 248 456 0 35 74 554 35 -1 554 24 74 248 -1