Approximate Numbers
Andrei has created his own game, which is quite easy to play. Each player must choose a playing card containing a string of n natural numbers. To win, the contestant must discover the fastest VALUE NEIGHBOR of each value in the string.

A value neighbor of a value x in a string is a number y with the property that y is the minimum of all numbers in the string greater than x.

To make sure he can win every time, Andrei asked you to create a program for him that determines the value neighbors of the n elements in the string of numbers on a playing card.

Input data
A natural number n, representing the number of elements on a playing card, is read from the keyboard, followed by a string of n natural numbers, representing the elements of the number string on that card.

Output data
The program will display on the screen a string of n natural numbers representing the string of value neighbors for each element x in the string. If an element in the read string has no value neighbor, "-1" will be displayed for it.

Restrictions and clarifications
0 < n ≤ 100
0 ≤ x ≤ 1000
Example
Input data Output data
10
15 47 5 12 21 5 64 85 64 12 21 64 12 15 47 12 85 -1 85 15