Couriers
Now that you've learned programming, you've started to notice all kinds of processes around you that could be optimized. For example, it is clear to you that if a courier has packages to deliver to 5 different houses, there is a way to order the 5 addresses so that the total time traveled by the courier is minimal.

Only one courier company operates in your hometown, Curier Rapid la Ușa Ta SRL. You offered to help them better organize their company couriers so they could deliver more packages, much faster. You have written a program that, given n addresses, calculates using Google maps the distance between any 2 addresses in the list. Then, you found out for each of the addresses which is the next address to which the courier can move the fastest, starting from it.

Now you have in front of you the result: A string of n distinct integer values, representing address indices, where the value at position i is the index of the house to which a courier must move after visiting the house at index i.

For example, for the string 2 3 4 1, the value at position 1 is 2, which means that if it starts at 1, it will continue to box 2. Try to complete the index string yourself before reading on!

He will go from 2 to 3, then from 3 to 4 and finally from 4 to 1, going through all the houses.

You noticed that in some cases multiple zones are created, where a courier starts from one point and ends up at the same point. For example, the string 2 3 1 5 4 has 2 zones: 1 2 3 and 4 5 (track a courier leaving from point 1, 2, 3, 4 or 5 to find out why).

Now write a program that receives such a string and displays the number of distinct zones in the string, so you can propose assigning one courier to each zone!

Input data
The number n will be found on the first line, and the elements of the string on the next line.

Output data
The program will display an integer on the screen, representing the number of zones formed by ordering the houses according to the read index string.

Specifications and restrictions
0 ≤ n ≤ 1000
the elements in the string will have distinct integer values ​​from 1 to n
Example
Input data Output data
9
4 2 9 1 3 7 8 6 5 4
Explanation
The zones are: (1, 4), (2), (3, 5, 9), (6, 7, 8)