Minimum Position
Andrei and Mihai want to apply for the same position in the WellCode team. Since the WellCode team is composed of elite programmers, only the best can get to be a part of it.

To recruit the best, Petru and Bianca asked the 2 candidates to write a program that, for a string of n ascending sorted natural numbers and a given value x, determines the minimum position p at which the value x is found in row.

The one whose program executes the task the fastest will enter the Wellcode team. Test your skills and solve the problem here on the platform 💪

Input data
The program reads from the keyboard, from the first line, the natural number n. From the next line, the string of n natural numbers, sorted in ascending order, is read. A natural number m is read from the third line, and m natural numbers are read from the fourth, representing possible values ​​for x.

Output data
It displays, for each of the m numbers read, the minimum position it is found in the original string. The values ​​will be displayed separated by spaces, and the numbering of the positions will be done starting with the value 1.

Restrictions and clarifications
1 ≤ n ≤ 500,000
the numbers in the string will have values ​​in the range [0, 500]
1 ≤ m ≤ 100,000
1 ≤ x ≤ 500
it is guaranteed that the values ​​of x read always exist in the given string
Examples
Input data Output data
10
1 2 3 4 4 4 5 6 6 7
4
2 4 5 6 2 4 7 8