Maximum Position
You finished the previous problem pretty quickly, well done! I modified the requirement a bit, can you solve this problem even faster than the first one?

This time, you are given an ascending string of n natural numbers and another string of m natural numbers. You have to find, for each of the m numbers, which is the rightmost position in the first string.

Input data
The program reads from the keyboard, from the first line, the natural number n. From the next line, the string of n natural numbers, sorted in ascending order, is read. A natural number m is read from the 3rd line, and m natural numbers are read from the 4th line, as required.

Output data
Display, for each of the m numbers read, the rightmost position it finds in the original string. The values ​​will be displayed separated by spaces, and the numbering of the positions will be done starting with the value 1.

Restrictions and clarifications
1 ≤ n ≤ 1 000 000
the numbers in the string will have values ​​in the range [1, 500]
1 ≤ m ≤ 1,000
1 ≤ x ≤ 500
it is guaranteed that all m values ​​read exist in the first string
Examples
Input data Output data
10
1 2 3 4 4 4 5 6 6 7
4
2 6 5 4 2 9 7 6