Circular permutations
Requirement
Consider a string consisting of n elements. We define the following operations as follows:

circular permutation left: move first element to end of string 1 2 3 4 -> 2 3 4 1 -> 3 4 1 2
circular permutation right: move the last element to the beginning of the string 1 2 3 4 -> 4 1 2 3 -> 3 4 1 2
Given a string consisting of n elements and two numbers k and p, to permute k positions to the right if p is -1, respectively to the left if p is 1.

Input data
On the first line there are 3 numbers: n , k and p. On the next line there are n natural numbers, representing the elements of the string.

Output data
n numbers will be displayed on a single line, separated by a space, representing the elements of the string obtained from the permutation operations.

limitation
1 ≤ n ≤ 1 000 000
0 ≤ k ≤ 1 000 000
The elements of the string are natural numbers between 1 and 1000
p can only have the values ​​1 and -1
Example
Input data Output data
4 1 -1
1 2 3 4 4 1 2 3
4 2 1
1 2 3 4 3 4 1 2