Search in an almost sorted string
Bogdan learned programming from WellCode and applied for a job at a large IT company. At the technical interview, he received the following problem:

A string of n natural numbers is given which has a more special property. Initially, it was sorted ascending, but an unknown number of left circular permutations were applied to it. k integer values ​​are also given. Find which of the k values ​​appear in the first string.

Bogdan did well in the interview and was hired! Now it's your turn to solve the problem.

Input data
The program reads from the first line a natural number n, and from the next line a string consisting of n natural numbers, having the properties from the statement. On the 3rd line a natural number k is read, and on the next line k integer values, representing the numbers you have to look for in the string.

Output data
The program displays on the screen, for each of the k values, if it is found or not in the given string. If x is found in the string, the message "x is found in the string" will be displayed. Otherwise, the message "x is not found in the string" will be displayed. (Where instead of x you will display the whole number sought)

This is an INTERVIEW PROBLEM you may encounter when applying for programming jobs!

Restrictions and clarifications
0 < n ≤ 50,000
the numbers in both strings have values ​​in the range (0, 1 000 000 000]
0 < k ≤ 10,000
Example
Input data Output data
7
5 12 15 17 20 2 4
4
1 15 5 7 1 is not found in the sequence
15 is found in the series
5 is found in the series
7 is not found in the series