EKG sequence
A sequence of integers can be graphed very easily if we add all the coordinate points (i, ai) on the graph. In this case, i represents a position of an element in the sequence, and ai represents the element at position i in the sequence.

For example, we can graph the sequence (1, 4, 3, 4) as follows:

In this problem you will use a sequence of integers discovered in 2001 and called the EKG Sequence, due to the similarity of its graph to an electrocardiogram. You can read more about her here.

The EKG sequence is defined as follows:

a[1] = 1
a[2] = 2
a[n] = the smallest number x that does not appear in previous positions in the sequence, and cmmdc(x, a[n - 1]) != 1, where cmmdc(a, b) is the greatest common divisor of a and b
Thus, the sequence starts with 1, 2, 4, 6, 3, 9, 12, 8, 10, 5, 15...

Requirement
Find the element at position n in the EKG sequence.

Input data
On the first line there is a single natural number, n.

Output data
A single natural number will be displayed, representing the element at position n in the EKG sequence.

limitation
1 ≤ n ≤ 1,000
Example
Input data Output data
10 5