File Sum
Alex recently organized a fundraising event for the National Informatics Batch. Each person participating in the event donated an amount of money. Alex wants to make a program that calculates how much money was raised at the event, knowing how many people attended and how much each donated.

Input data
The program reads from the file donatii.in a number n, representing the number of people participating in the event, and from the next line n integer values, representing the amounts of money donated by each person.

Output data
The program will display on the screen the total amount of money collected at the event.

Restrictions and clarifications
each person donates an amount of money in the range [0, 1 000].
0 < n ≤ 100
CAREFUL! Because of the way this problem is evaluated, when loading the solution to the platform you will only need to add the int main() statements, and at the end it is very important NOT to add return 0;.
all the libraries you need are already included in the program
Example
Input data Output data
13
5 15 258 0 17 64 99 201 784 963 10 574 248 3238