Deleting a row from an array
Given a matrix with n rows and m columns, and a natural number k. Delete row k from the matrix.

Input data
On the first line, the numbers n, m, and k are read from the keyboard, respectively, with the meanings in the statement. m natural numbers are read on the next n lines.

Output data
The program will display the matrix with the kth row deleted on the screen.

limitation
1 < k <= n, m < 500

Example
Input data Output data
3 3 2
1 2 3
4 5 6
7 8 9 1 2 3
7 8 9