Quadratic matricesParallel to the main diagonal
Given a matrix m with n rows and n columns, and two numbers i and j representing the matrix position of the first element of a parallel to the main diagonal. Display all elements from top to bottom on that parallel to the diagonal.

Input data
From the first line, the number n is read from the keyboard. From the next n lines, n natural numbers are read that represent the elements of the matrix. From the last line the numbers i and j are read.

Output data
The program will display on the screen a string of numbers separated by spaces, representing the elements from the parallel to the diagonal.

Restrictions and clarifications
1 ≤ n ≤ 50
0 < i, j ≤ n
1 ≤ m[i][j] ≤ 100
always i = 1 or j = 1
rows and columns are numbered from 1 to n
Example
Input data Output data
4
1 2 3 4
5 6 7 8
9 10 11 12
13 14 15 16
1 2 2 7 12
4
1 2 3 4
5 6 7 8
9 10 11 12
13 14 15 16
2 1 5 10 15