Parallel to the main diagonal
Given a matrix m with n rows and n columns. Show each parallel to the main diagonal. The parallels will be traveled from left to right. The elements on each parallel will be traversed from bottom to top.

Input data
From the first line, the number n is read from the keyboard. From the next n lines, n natural numbers are read that represent the elements of the matrix.

Output data
The program will display on the screen a string of numbers, separated by spaces, representing the elements of the matrix traversed in the order of the statement.

Restrictions and clarifications
1 ≤ n ≤ 50
1 ≤ m[i][j] ≤ 100
Example
Input data Output data
4
1 2 3 4
5 6 7 8
9 10 11 12
13 14 15 16 13 14 9 15 10 5 16 11 6 1 12 7 2 8 3 4