Rotate an array to the left
Given a matrix mt with n rows and m columns. To rotate the given matrix by 90 degrees to the left.

Input data
From the first line, the numbers n and m are read from the keyboard. From the next n lines, m natural numbers are read that represent the elements of the matrix.

Output data
The program will display the matrix resulting from the rotation on the screen.

Restrictions and clarifications
1 ≤ n, m ≤ 50
1 ≤ mt[i][j] ≤ 3000
Example
Input data Output data
4 3
1 2 3
5 6 7
9 10 11
13 14 15 3 7 11 15
2 6 10 14
1 5 9 13