Image intensity
Story
You have just been hired in the image processing department of a large company. To get used to your new job, you were given a simple task: You will receive a black-and-white image and you must find out for various areas in the image what the total white intensity is in that area.

The image you get is in the form of an mt matrix with n rows and m columns. Each element i, j of the matrix represents the intensity of the pixel on row i and column j in the image.

Your manager gives you g zones in the image, and you have to tell him for each of the g zones what the total intensity of the zone is. The total intensity of an area is the sum of the intensities of each pixel in the area.

An area in the image is defined by the numbers i1, j1, i2, j2. The area is actually the submatrix with the upper-left (i1, j1) and lower-right (i2, j2) corners.

Requirement
Given the image and the g zones, find the total intensity for each zone.

Input data
From the first line, the numbers n and m are read from the keyboard. From the next n lines, m natural numbers are read that represent the intensity of each pixel.

The next line will read the number g, representing the number of areas in the image for which you need to find the total intensity.

On each of the following g lines there will be 4 numbers i1, j1, i2, j2, representing the coordinates of the submatrix.

Output data
The program will display g lines on the screen, representing the total intensity of each zone.

limitation
1 ≤ n, m ≤ 100
-3 000 ≤ mt[i][j] ≤ 3 000
1 ≤ g ≤ 100
1 ≤ i1 ≤ i2 ≤ n
1 ≤ j1 ≤ j2 ≤ m
Example
Input data Output data
4 5
-1 -1 -4 -4 -1
-1 -1 -5 -5 -1
-1 -1 -1 -9 -1
-1 -6 -7 -8 -1
3
1 1 4 5
3 4 3 4
2 2 3 4 -60
-9
-22
Explanation
For the submatrix (1, 1), (4, 5), its intensity is equal to the sum of all elements in the matrix.

The submatrix (3, 4), (3, 4) consists only of the element on row 3, column 4.

The submatrix (2, 2), (3, 4) is highlighted below:

-1
-1
-1
-1
-1
-1
-1
-1
-1
-1
-1
-1
-6
-7
-4
-5
-9
-8
-4
-5