SumColMax
Requirement
Dorel, farmer and chief, has an agricultural land of n x m km2 on which he regularly grows wheat. Now is the harvest season, and on every km2 of land there is a (known) amount of wheat.

Dorel's farm has m employees, whom he distributes so that each one gathers the wheat from a single column. The amount that a worker collects is the sum of the values ​​of the elements located on the column where he is placed.

A matrix consisting of n lines and m columns is given, representing the scheme of the agricultural land. Display the quantities of wheat from each km2 that are gathered by the most diligent worker. The most diligent worker is the one who gathered the most wheat.

Input data
On the first line, read 2 natural numbers, n and m, separated by a space. On the next n lines there are m integers separated by a space, the elements of the matrix representing the agricultural land.

Output data
n numbers separated by a space will be displayed, representing the elements of the column from which the most diligent worker collects. These numbers will be displayed in the order they were read.

limitation
1 ≤ n, m ≤ 1,000
Array elements will be integers greater than -1000 and less than 1000
If there are several workers collecting the maximum amount, only the one with the lowest column index will be considered
Paradoxically, the amount of wheat a worker gathers can be a negative number
Example
Input data Output data
4 5
2 4 2 6 7
1 3 1 2 5
5 3 2 5 6
4 4 5 2 1 7 5 6 1
3 3
1 2 3
2 3 1
3 1 2
1 2 3