Zones
A natural number n is given. Generate a matrix with n lines and n columns, in which the main and secondary diagonals are 0 and then the 4 areas formed by the 2 diagonals are coded with 1, 2, 3, 4, as in the example.

Input data
The natural number n is read from the keyboard.

Output data
The program will display the required matrix consisting of n rows and n columns. There will be a space between the numbers on a line.

limitation
1 ≤ n ≤ 50
Example
Input data Output data
5 0 1 1 1 0
4 0 1 0 2
4 4 0 2 2
4 0 3 0 2
0 3 3 3 0