Move zones
Given a square matrix of size n. It is divided into four sectors by the main and minor diagonals. To carry out a circular permutation of the sectors, clockwise.

Input data
On the first line is the number n. On the next n lines are n numbers separated by a space, representing the elements of the matrix.

Output data
The program will display the matrix after performing the permutation.

limitation
1 ≤ n ≤ 50
1 ≤ m[i][j] ≤ 100
Examples
Input data Output data
3
1 2 3
4 5 6
7 8 9 1 4 3
8 5 2
7 6 9
4
11 12 13 14
21 22 23 24
31 32 33 34
41 42 43 44 11 31 21 14
42 22 23 12
43 32 33 13
41 34 24 44
5
11 12 13 14 15
21 22 23 24 25
31 32 33 34 35
41 42 43 44 45
51 52 53 54 55 11 41 31 21 15
52 22 32 24 12
53 43 33 23 13
54 42 34 44 14
51 45 35 25 55